Feat: Add form input validation hook
