+++ app/_hooks/useFormInput.ts
@@ -1,0 +1,45 @@
+import { validate, IErrors } from '@/app/_lib/validate'
+import { ChangeEvent, useState } from 'react'
+
+interface IInputFileds {
+  email?: string
+  emailCheck?: string
+  password?: string
+  passwordCheck?: string
+}
+
+export default function useFormInput(initialValues: IInputFileds) {
+  const [inputFields, setInputFields] = useState<IInputFileds>({
+    ...initialValues,
+  })
+  const [errors, setErrors] = useState<IErrors>({})
+  const [submitting, setSubmitting] = useState(false)
+
+  const handleChange = (e: ChangeEvent<HTMLInputElement>) => {
+    const { name, value } = e.target
+
+    const updatedInputFields = { ...inputFields, [name]: value }
+    setInputFields(updatedInputFields)
+
+    const validationErrors = validate(updatedInputFields)
+    setErrors(validationErrors)
+  }
+
+  const handleSubmit = (e: ChangeEvent<HTMLFormElement>) => {
+    e.preventDefault()
+    setSubmitting(true)
+  }
+
+  const finishSubmit = () => {
+    console.log(inputFields)
+  }
+
+  return {
+    inputFields,
+    errors,
+    submitting,
+    handleSubmit,
+    handleChange,
+    finishSubmit,
+  }
+}
