+++ app/_lib/validate.ts
@@ -1,8 +1,8 @@
 interface IInputValues {
-  email: string
-  password: string
-  passwordCheck: string
-  emailCheck: number | null
+  email?: string
+  password?: string
+  passwordCheck?: string
+  emailCheck?: string
 }
 export interface IErrors {
   email?: string
@@ -13,16 +13,20 @@ export interface IErrors {
 
 export const validate = (inputValues: IInputValues): IErrors => {
   const errors: IErrors = {}
-  if (inputValues.email.length < 6) {
+  if (inputValues.email && inputValues.email.length < 6) {
     errors.email = '이메일 길이를 6자 이상 적어주세요.'
   }
-  if (inputValues.password.length < 6) {
+  if (inputValues.password && inputValues.password.length < 6) {
     errors.password = '비밀번호 길이를 6자 이상 적어주세요.'
   }
-  if (inputValues.password !== inputValues.passwordCheck) {
+  if (
+    inputValues.password &&
+    inputValues.passwordCheck &&
+    inputValues.password !== inputValues.passwordCheck
+  ) {
     errors.passwordCheck = '비밀번호가 일치하지 않습니다.'
   }
-  if (inputValues.emailCheck === null || inputValues.emailCheck !== 1234) {
+  if (inputValues.emailCheck && inputValues.emailCheck !== '1234') {
     errors.emailCheck = '인증번호가 일치하지 않습니다.'
   }
   return errors
